# Optimize parts of the entity cramming check