--- a/src/main/java/me/jellysquid/mods/lithium/common/config/LithiumConfig.java
+++ b/src/main/java/me/jellysquid/mods/lithium/common/config/LithiumConfig.java
@@ -77,6 +77,7 @@ public class LithiumConfig {
 
         this.addMixinRule("entity", true);
         this.addMixinRule("entity.cached_climbing_check", true);
+        this.addMixinRule("entity.cached_health", true);
         this.addMixinRule("entity.collisions", true);
         this.addMixinRule("entity.data_tracker", true);
         this.addMixinRule("entity.data_tracker.no_locks", true);