--- a/src/main/java/me/jellysquid/mods/lithium/common/config/LithiumConfig.java
+++ b/src/main/java/me/jellysquid/mods/lithium/common/config/LithiumConfig.java
@@ -76,6 +76,7 @@ public class LithiumConfig {
         this.addMixinRule("collections.entity_filtering", true);
 
         this.addMixinRule("entity", true);
+        this.addMixinRule("entity.cached_climbing_check", true);
         this.addMixinRule("entity.collisions", true);
         this.addMixinRule("entity.data_tracker", true);
         this.addMixinRule("entity.data_tracker.no_locks", true);