--- a/src/main/resources/lithium.mixins.json
+++ b/src/main/resources/lithium.mixins.json
@@ -92,6 +92,7 @@
     "chunk.serialization.PackedIntegerArrayMixin",
     "chunk.serialization.PalettedContainerMixin",
     "collections.entity_filtering.TypeFilterableListMixin",
+    "entity.cached_climbing_check.MixinLivingEntity",
     "entity.collisions.CollisionViewMixin",
     "entity.collisions.EntityMixin",
     "entity.collisions.EntityViewMixin",