--- a/src/main/resources/lithium.mixins.json
+++ b/src/main/resources/lithium.mixins.json
@@ -93,6 +93,7 @@
     "chunk.serialization.PalettedContainerMixin",
     "collections.entity_filtering.TypeFilterableListMixin",
     "entity.cached_climbing_check.MixinLivingEntity",
+    "entity.cached_health.MixinLivingEntity",
     "entity.collisions.CollisionViewMixin",
     "entity.collisions.EntityMixin",
     "entity.collisions.EntityViewMixin",