Some preliminary results:
```
All tests using 365 chickens on a ladder in an empty world after a fresh restart and a 3600-tick warp.

Pre-patch:
	Trial 1:
		Total:		12.565ms
		Chicken:	11.77ms
	Trial 2:
		Total:		12.235ms
		Chicken:	11.45ms
	Trial 3:
		Total:		12.432ms
		Chicken:	11.63ms
	Average:
		Total:		12.411ms
		Chicken:	11.62ms

a4f0ef38:
	Trial 1:
		Total:		9.613ms
		Chicken:	8.56ms
	Trial 2:
		Total:		9.299ms
		Chicken:	8.35ms
	Trial 3:
		Total:		9.312ms
		Chicken:	8.38ms
	Average:
		Total:		9.408ms
		Chicken:	8.43ms

f97444b6:
	Trial 1:
		Total:		7.781ms
		Chicken:	6.72ms
	Trial 2:
		Total:		7.605ms
		Chicken:	6.62ms
	Trial 3:
		Total:		7.478ms
		Chicken:	6.57ms
	Average:
		Total:		7.621ms
		Chicken:	6.64ms

ba1b3e48:
	Trial 1:
		Total:		8.251ms
		Chicken:	7.25ms
	Trial 2:
		Total:		8.153ms
		Chicken:	7.22ms
	Trial 3:
		Total:		8.158ms
		Chicken:	7.24ms
	Average:
		Total:		8.187ms
		Chicken:	7.24ms

525609dd:
	Trial 1:
		Total:		7.652ms
		Chicken:	6.70ms
	Trial 2:
		Total:		7.565ms
		Chicken:	6.69ms
	Trial 3:
		Total:		7.505ms
		Chicken:	6.61ms
	Average:
		Total:		7.574ms
		Chicken:	6.67ms

5589b3d5:
	Trial 1:
		Total:		7.139ms
		Chicken:	6.11ms
	Trial 2:
		Total:		6.961ms
		Chicken:	6.04ms
	Trial 3:
		Total:		6.876ms
		Chicken:	5.96ms
	Average:
		Total:		6.992ms
		Chicken:	6.04ms

Average Improvement:
	a4f0ef38:
		Total:		-3.003ms	-24.2%
		Chicken:	-3.19ms		-27.5%
	f97444b6:
		Total:		-4.790ms	-38.6%
		Chicken:	-4.98ms		-42.9%
	ba1b3e48:
		Total:		-4.224ms	-34.0%
		Chicken:	-4.38ms		-37.7%
	525609dd:
		Total:		-4.837ms	-39.0%
		Chicken:	-4.95ms		-42.6%
	5589b3d5:
		Total:		-5.419ms	-43.7%
		Chicken:	-5.58ms		-48.0%

```