@magneticflux- 2021-07-26
I think this is as optimized as it gets without rewriting the collision system to act per-world (using some kind of disjoint-set or octree) instead of per-entity.