@magneticflux- 2021-08-22
Yes, I believe this still provides a statistically significant improvement in TPS while not obviously breaking vanilla parity like the Carpet mod does.