@2No2Name 2021-08-22
I believe this PR breaks vanilla parity in cases where entities collide with climbable blocks that are placed or removed. Therefore this cannot be merged