@2No2Name 2021-08-22
After a discussion on discord we agreed on trying to add better cache invalidation to this PR to resolve the problems.
1) Invalidate when the entity moves
2) Invalidate when the chunk changes