@magneticflux- 2021-08-22
Didn't have the original setup, but I redid some of it from scratch:
```
All tests using 512 chickens on a ladder in an empty world after a fresh restart and a 3600-tick warp.

Pre-patch:
	Trial 1:
		Total:		25.331ms
		Chicken:	22.84ms
	Trial 2:
		Total:		25.698ms
		Chicken:	23.14ms
	Trial 3:
		Total:		25.380ms
		Chicken:	22.85ms
	Average:
		Total:		25.470ms
		Chicken:	22.94ms

a69f8082:
	Trial 1:
		Total:		17.647ms
		Chicken:	14.42ms
	Trial 2:
		Total:		17.430ms
		Chicken:	14.34ms
	Trial 3:
		Total:		17.030ms
		Chicken:	14.21ms
	Average:
		Total:		17.369ms
		Chicken:	14.32ms

Average Improvement:
	a69f8082:
		Total:		-8.101ms	-31.8%
		Chicken:	-8.62ms		-37.5%
```