@2No2Name 2021-08-23
Can confirm, blinking redstone lines went from 41mspt to 31mspt.