@2No2Name 2021-08-23
I will profile the isClimbing caching. The health caching just seems to duplicate the information from the datatracker to avoid the datatracker. If that turns out to be a good idea, I would like to find out whether that can be applied to other tracked data as well (e.g. item stack of item entities).