@2No2Name 2021-08-23
I just noticed that the ladder system still has some flaws:

- LivingEntity.setPosition is not the only place where the position is modified
- Trapdoors on top of ladders are like ladders only if there is a ladder below. The ladder below can be in a different chunk section, and therefore the cache should be invalidated when the ladder is broken, but isn't.