@2No2Name 2021-08-23
![2021-08-23_19 20 58](https://user-images.githubusercontent.com/50278648/130490297-fc18029c-20d4-41b7-a7ce-d933f174ac08.png) shows a setup that demonstrates the problem. Climbing up, then breaking the ladders below, will lead to the player still climbing the trapdoors, even though they are no longer climbable. Same test without the patch will cause the player to drop when the ladder is broken