@2No2Name 2021-08-23
Pushed my changes here. https://github.com/2No2Name/lithium-fabric/tree/feature/entity-cramming The performance is less than 10% better in my testing